now using TableVC
